Adding extra logging and safety for hung child processes

+++ src/xpc/execute_process_queue.ts
@@ -9,18 +9,34 @@ export class ExecuteProcessQueue extends ProcessQueue {
             return new Promise<string>((resolve, reject) => {
                 const child = spawn.fork(`./src/xpc/execute_process.ts`)
                 child.on("message", (actionResponse) => {
-                    child.kill()
                     resolve(actionResponse)
+                    child.kill()
                 }).on("error", (err) => {
-                    winston.warn(err.message)
+                    winston.warn(`ChildProcess sent error message: ${err.message}`)
                     if (!child.killed) {
                         child.kill()
                     }
                     reject(err)
-                }).on("exit", () => {
+                }).on("exit", (code: number, signal: string) => {
+                    winston.info(`ChildProcess exited with code: ${code}, signal: ${signal}`)
                     if (!child.killed) {
                         child.kill()
                     }
+                    reject(signal)
+                }).on("disconnect", () => {
+                    winston.warn(`ChildProcess disconnected`)
+                    if (!child.killed) {
+                        child.kill()
+                    }
+                    reject("Child Disconnected")
+                }).on("close", (code: number, signal: string) => {
+                    if (signal !== "SIGTERM") {
+                        winston.warn(`ChildProcess closed with code: ${code}, signal: ${signal}`)
+                    }
+                    if (!child.killed) {
+                        child.kill()
+                    }
+                    reject(signal)
                 })
                 child.send(data)
             })
